c file

+++ 0x06-pointers_arrays_strings/103-infinite_add.c
@@ -34,7 +34,7 @@ char *infinite_add(char *n1, char *n2, char *r, int size_r)
 				r[j] = n1[j];
 			else
 			{
-				res = res + (int) (n1[j] - '0' + n2[j] - '0');
+				res = res + n1[j] - '0' + n2[j] - '0';
 				r[j] = (res % 10) + '0';
 				res = res / 10;
 			}
